backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
FastAPI==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
amadeus==8.1.0
googlemaps==4.10.0
python-dateutil==2.8.2
geopy==2.4.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
asyncio==3.4.3 